requirements.txt
streamlit
numpy
pandas
matplotlib
yfinance
scikit-learn
Pillow
requests
streamlit-lottie
